Fix project reference paths
Previously the project files were referring to some assembly references
via relative paths (from the project) to the fixed Program Files x86
folder.

Trouble is that the relative path from the project file to the Program
Files folder depends on the repository root that you cloned into.

By qualifying the assemblies that are descendants Program Files off of
another path that is already in Program Files, this should make it
possible to build the solution regardless of which path you clone into.